Improve render performance of CodeMirror
On WE2010 we were telling CodeMirror to draw all lines. This was in
order to make sure that browser search could still detect all the
content (T174480). This comes with significant performance problems.

Instead hook into the events of CodeMirror and only set this huge
drawsurface when there haven't been changes for the last second.

This will make the editor feel more responsive on large pages, while
still allowing to use browsersearch when you are not making changes at
the same time.

Bug: T197502
Change-Id: Iefab764672cfe16397421d2b393006086f4c2bfc